add base module
