+++ facebook/fb.py
@@ -4,3 +4,10 @@ from graph import GraphAPIError, get_graph, get_static_graph, get_public_graph
 from session import get_session
 from modules.profile.application.utils import get_app_dict
 from fql import get_FQL
+
+from facebook.modules.profile.user.models import User, TestUser
+from facebook.modules.profile.page.models import Page
+from facebook.modules.profile.event.models import Event
+from facebook.modules.profile.application.models import Request, Score
+from facebook.modules.media.models import Photo
+from facebook.modules.connections.post.models import Post, PostBase
