+++ facebook/modules/media/models.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from django.db import models
-from facebook.models import Base
+from facebook.modules.base import Base
 from facebook.modules.profile.user.models import User
 from facebook.graph import get_graph
 from django.utils.translation import ugettext_lazy as _
