+++ facebook/modules/profile/models.py
@@ -2,7 +2,7 @@ from django.db import models
 from django.template.defaultfilters import slugify
 from django.utils.translation import ugettext_lazy as _
 
-from facebook.models import Base
+from facebook.modules.base import Base
 
 
 class Profile(Base):
