+++ facebook/modules/profile/user/admin.py
@@ -1,12 +1,7 @@
-from django.contrib import admin
 
-from facebook.admin import AdminBase
-
-from .models import User
-
+from facebook.modules.base import AdminBase
 
 class UserAdmin(AdminBase):
     list_display = ('id', 'profile_link', 'access_token', 'user', '_name', 'created', 'updated',)
     readonly_fields = ('friends', '_name', '_first_name', '_last_name', '_link', '_birthday', '_email', '_location', '_gender', '_graph')
     search_fields = ('id', '_name')
-admin.site.register(User, UserAdmin)
